Finished the debug log scheme. Note that you can use the public constructor that takes no args and it will create a debug log of the form DBoutN.txt where "N" is the number of log created.
Otherwise, you can specify a filename if you want to use specific names
to categorize error output. You can also just use the Debug.Log() method
to print things to the console. This functions the same as
System.out.println().

In the case of instantiating objects to write to files...you need to
close the log once you're done with close(). Otherwise nothing will
print to the file.