Removed ScreenFactory.java because it was redundant and inefficient
Deleted redundant code related to the ScreenFactory class

Renamed SampleGameScreen.java to GameScreen, and will eventually be
using this as a camera class (I think)

The fixedUpdate() function is finally in place in the game loop.

Deleted the Input.java file because it's not necessary yet, and I don't
know how to implement static input. This would be nice to figure out.

Probably did some other things. I don't remember what at the moment.