Major change to the physics system. From top down:
- GameEntity: The updatePhysics() method is re-written. Now, the entire
rigidbody associated with the GameEntity is passed to the Physics class
for integration. This way, we can keep track of all important properties
of the RigidBody (currently position, velocity, mass, and drag
coefficient) without adding extra parameters to the Physics integration
functions. Adding further RigidBody parameters is much more feasible
now.

- Physics: Re-written to accept just a desired timestep and a RigidBody
object. All other functionality is internal now. The integration methods
were re-written also. RK4 is still the only technique available, but I
will start adding more methods once I can think of a good way to perform
a "verifiable" integration with an analytical solution.

- PlayerController: Modified to use a force instead of operating
directly on the velocity of the object. This has a different control
"feel," but allows easier consideration of things such as mass and
damping. For objects requiring a quicker, more "gamey" control response,
we'll devise a kinematic movement system which circumvents most of the
physics integration and responds directly to control inputs.

Main: Added a debug flag to allow test methods to print to the console.
I'll probably use this extensively as I move forward with verifying the
correctness of our integration techniques.