Major refactoring/organization update: Created a physics package Created a graphics package
Ideally this reorganization will lend itself to a more logical project
structure over time, as related classes can be grouped into these
packages, and low level engine code can be limited to package-access
methods.

Collider:
-All colliders should now be created using the Collider.createCollider
method
Renderer:
-All renderers should now be created using the Renderer.createRenderer
method

The function of these methods is to begin isolation of the creation of a
component from the implementation of it. Certain types of components
that are supported will be added to the enums behind these methods.